feat(run_all): run_all gathering all test files for umbrella apps
For now, this uses Mix.Project.apps_paths() and expects all test files
to be in a "test" dir - it'd be nice to use each sub-app's
config[:test_paths], but we may never need it, as the rest of the test
files will have already been loaded anyway. Happy to dig further in when
that time comes. This also sets an infinite timeout on the genserver
call, as an umbrella's worth of tests won't always beat 5 seconds.