feat(file_watcher): adds dependency directories to the file watcher
Dependencies will now re-compile when saved as well as normal source
code. This is useful when you are using local path: "../project_dir",
and want to make changes to both the current app and the dependencies as
well. This also allows you to edit your application's deps while the app
is running, which can be useful when debugging + interacting with new
dependencies.