Fix Code.load_file deprecation
Code.load_file was deprecated in Elixir
1.10 (https://hexdocs.pm/elixir/compatibility-and-deprecations.html)

This replaces it with `Code.compile_file/1` (`Code.require_file/1` was not used
because we want to be able to recompile the same file multiple times).

Since `Code.compile_file/1` was added in Elixir 1.7, support for 1.6 was
removed. 1.7 was released in July 2018 so it is almost 2 years old at this
point, so hopefully it is okay to drop support.

https://elixir-lang.org/blog/2018/07/25/elixir-v1-7-0-released/